handeled NULL string

+++ main.h
@@ -3,7 +3,7 @@
 
 int _putchar(char c);
 int _putstr(char* str);
-int _putnum(int num);
+int _putnum(long int num);
 int _printf(const char *format, ...);
 
 #endif
